style: 메인 페이지 스타일 수정
