+++ src/components/Header/index.jsx
@@ -1,22 +1,26 @@
 import React from 'react';
-import { StHeader } from './styles';
+import { NavLink } from 'react-router-dom';
 import { ReactComponent as Logo } from 'assets/logo.svg';
-import { Link } from 'react-router-dom';
+import { StHeader } from './styles';
 
 export default function Header() {
   return (
     <StHeader>
       <nav>
         <ul>
-          <Link to="/">
-            <li>home</li>
-          </Link>
-          <li>member</li>
+          <li>
+            <NavLink
+              to="/"
+              className={({ isActive }) => (isActive ? 'active' : '')}
+            >
+              home
+            </NavLink>
+          </li>
         </ul>
         <h1>
           <Logo fill="white" width="200" height="70" stroke="white" />
         </h1>
-        <ul>{/* <li>about</li> */}</ul>
+        <ul></ul>
       </nav>
     </StHeader>
   );
