+++ src/components/Letter/index.jsx
@@ -13,7 +13,7 @@ export default function Letter({ memberId, messages }) {
 
   return (
     <S.Letter>
-      <S.H3>To. {memberData.englishName}</S.H3>
+      <h3>To. {memberData.englishName}</h3>
       {filterdData?.map((item) => {
         const { nickname, avatar, content, writedTo, createdAt, id } = item;
         return (
