+++ src/components/LetterItem/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import * as S from './styles';
 import { Link } from 'react-router-dom';
+import { timestampToDate } from 'utils/date';
 
 export default function LetterItem({
   nickname,
@@ -12,27 +13,22 @@ export default function LetterItem({
 }) {
   return (
     <S.Letter>
-      <Link
-        to={{
-          pathname: `detail/${id}`,
-          state: {
-            nickname,
-            avatar,
-            content,
-            createdAt,
-            writedTo,
-            id,
-          },
-        }}
-      >
-        <img src={!avatar ? require(`assets/default.jpg`) : avatar} alt="" />
-        <div>
-          <h4>{nickname}</h4>
+      <Link to={`detail/${id}`}>
+        <S.LetterImg>
+          <img
+            src={!avatar ? require(`assets/defaultUser.jpg`) : avatar}
+            alt=""
+          />
+        </S.LetterImg>
+        <S.LetterInfo>
           <div>
-            <span title={content}>{content}</span>
-            <h5>{new Date(createdAt).toLocaleString().substring(2, 21)}</h5>
+            <span>{nickname}</span>
+            <span>{timestampToDate(createdAt)}</span>
           </div>
-        </div>
+          <div>
+            <p title={content}>{content}</p>
+          </div>
+        </S.LetterInfo>
       </Link>
     </S.Letter>
   );
