+++ src/components/MemberList/index.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { StPolaroid } from './styles';
+import * as S from './styles';
 
 export default function MemberList({ children: name, id, clicked, onUpdate }) {
   const selectMember = () => onUpdate(id);
 
   return (
-    <StPolaroid onClick={selectMember} $clicked={clicked}>
+    <S.MemberList onClick={selectMember} $clicked={clicked}>
       <div>
         <img src={require(`assets/${name}.jpg`)} alt={name} />
         <p>{name}</p>
       </div>
-    </StPolaroid>
+    </S.MemberList>
   );
 }
